feat: implement deletion of branches that track a remote branch
This is a first attempt implementation of an application that will clean
up local git branches that have been deleted from a remote repository.
This is done by looking at the local branches and making a list of all
of them that track a given remote. The remote is queried to see what
branches it has. If a local branch is tracking that remote but the
remote doesn't have that branch then the branch will be deleted.

The `-d` flag will perform a dry-run and return the list of branches
that will be deleted.

By default the application will not perform a dry run and will use the
remote `origin`.